requirements.txt
PySide2~=5.14.2.1
python-dateutil~=2.8.1
requests~=2.26.0